Update mortality() absorption decompositions
Switch from decomposing total result into individual components to
calculating individual components directly. This removes an
multiplication and division for the individual components that lead to
reduced precision. Performance may also benefit, but it's not clear if
that is the case

Make sure result elements are all named correctly/consistently