Update net visitation calculation
Change the net visitation function from the old approach, which
calculates the flow on edges between nodes i and j and then splits it
between the two nodes, to assigning the flow to just node j. Then add
the initial state vector to the result to account for starting flow.
This properly handles flow in the generic case of multiple starting
states and multiple absorbing states, unlike the gdistance
implementation that assumes a single starting node and single fully
absorbing node.